подчистил features исправил нейминг вынес composable компоненты в common.ui вынес delay в SharingStarted.WhileSubscribed в константы вынес extension функции в common.extensions вынес enum классы
