+++ app/src/main/java/com/example/bankapp/core/navigation/HistoryType.kt
@@ -1,0 +1,5 @@
+package com.example.bankapp.core.navigation
+
+enum class HistoryType {
+    INCOME, EXPENSE
+}
