+++ app/src/main/java/com/example/bankapp/features/account/AccountsScreen.kt
@@ -3,7 +3,6 @@ package com.example.bankapp.features.account
 import ListItem
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -19,12 +18,12 @@ import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.example.bankapp.R
-import com.example.bankapp.features.common.CurrencyChangeBottomSheet
-import com.example.bankapp.features.common.LazyList
-import com.example.bankapp.features.common.LeadIcon
-import com.example.bankapp.features.common.PriceDisplay
-import com.example.bankapp.features.common.ResultStateHandler
-import com.example.bankapp.features.common.TrailingContent
+import com.example.bankapp.features.common.ui.CurrencyChangeBottomSheet
+import com.example.bankapp.features.common.ui.LazyList
+import com.example.bankapp.features.common.ui.LeadIcon
+import com.example.bankapp.features.common.ui.PriceDisplay
+import com.example.bankapp.features.common.ui.ResultStateHandler
+import com.example.bankapp.features.common.ui.TrailingContent
 
 
 
@@ -33,11 +32,11 @@ fun AccountsScreen(
     viewModel: AccountViewModel = hiltViewModel()
 ) {
 
-    val mock by viewModel.accountState.collectAsStateWithLifecycle()
+    val state by viewModel.accountState.collectAsStateWithLifecycle()
     var showBottomSheet by remember { mutableStateOf(false) }
 
     ResultStateHandler(
-        state = mock,
+        state = state,
         onSuccess = { data ->
             if (showBottomSheet) {
                 CurrencyChangeBottomSheet({showBottomSheet = false})
