--- app/src/main/java/com/example/bankapp/features/common/date/DateMode.kt
@@ -1,5 +1,0 @@
-package com.example.bankapp.features.common.date
-
-enum class DateMode {
-    START, END
-}
