--- app/src/main/java/com/example/bankapp/features/common/ui.kt
@@ -1,2 +1,0 @@
-package com.example.bankapp.features.common
-
