+++ app/src/main/java/com/example/bankapp/features/common/ui/AddButton.kt
@@ -1,4 +1,4 @@
-package com.example.bankapp.features.common
+package com.example.bankapp.features.common.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
