+++ app/src/main/java/com/example/bankapp/features/expenses/ExpensesViewModel.kt
@@ -2,11 +2,12 @@ package com.example.bankapp.features.expenses
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.bankapp.domain.model.IncomeInfo
 import com.example.bankapp.domain.model.Transaction
 import com.example.bankapp.domain.repository.AccountRepository
 import com.example.bankapp.domain.repository.TodayTransactionRepository
 import com.example.bankapp.core.ResultState
+import com.example.bankapp.features.common.extensions.filterExpenses
+import com.example.bankapp.utils.Delays
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
@@ -28,14 +29,14 @@ class ExpensesViewModel @Inject constructor(
         }
         .stateIn(
             viewModelScope,
-            SharingStarted.WhileSubscribed(5000L),
+            SharingStarted.WhileSubscribed(Delays.STOP_TIMEOUT_MILES),
             ResultState.Loading)
 
     private var _totalExpensesState = MutableStateFlow<Double>(0.0)
     val totalExpensesState = _totalExpensesState
         .stateIn(
             viewModelScope,
-            SharingStarted.WhileSubscribed(5000L),
+            SharingStarted.WhileSubscribed(Delays.STOP_TIMEOUT_MILES),
             0.0)
 
     private fun loadExpensesTransactions(){
@@ -61,9 +62,3 @@ class ExpensesViewModel @Inject constructor(
     }
 }
 
-fun<T: IncomeInfo> ResultState<List<T>>.filterExpenses(): ResultState<List<T>> =
-
-    when (this) {
-        is ResultState.Success -> ResultState.Success(this.data.filter { !it.isIncome })
-        else -> this
-    }
