+++ app/src/main/java/com/example/bankapp/features/history/models/DateMode.kt
@@ -1,0 +1,5 @@
+package com.example.bankapp.features.history.models
+
+enum class DateMode {
+    START, END
+}
