+++ app/src/main/java/com/example/bankapp/features/history/models/DatePickerState.kt
@@ -1,0 +1,7 @@
+package com.example.bankapp.features.history.models
+
+enum class DatePickerState {
+    CLOSED,
+    OPEN_START,
+    OPEN_END
+}
