+++ app/src/main/java/com/example/bankapp/features/settings/SettingsScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.example.bankapp.R
-import com.example.bankapp.features.common.LazyList
+import com.example.bankapp.features.common.ui.LazyList
 import com.example.bankapp.features.settings.utils.SettingsItems
