+++ app/src/main/java/com/example/bankapp/utils/Constants.kt
@@ -3,4 +3,5 @@ package com.example.bankapp.utils
 object Delays {
     const val ERROR_500_RETRY = 2000L
     const val INTERNET_ERROR_RETRY = 7000L
+    const val STOP_TIMEOUT_MILES = 5000L
 }
